Add ability to embed additional content in the diff view
In downstream applications we may want to add extra data to our scroll
view. With the way the widget layout has to be done, it's hard to
decouple the DiffView widget from the search widget without forcing
downstream to re-implement some unnecessary logic.

Provide a callback parameter that allows a caller to inject whatever
they want (e.g. a commit message) into the diff view before the diff is
rendered