Decouple DiffView from ScrollArea
For external users (sphaerophoria/spit), we may want to embed a DiffView
in our own ScrollArea. We may want to embed multiple DiffViews in the
same ScrollArea (e.g. Index + Working tree in a git gui). With the
current implementation we are unable to do that because the DiffView
will expand to fill up all remaining space within the given ScrollArea

* Move SearchBar ownership to top level, outside the diff view
* Add wrapper function to append a UI with a SearchBar. This does not
  necessarily have to be a scroll area, but should support any widget
  that expands to fill the full remaining space
* Remove unnecessary function to embed data within our internal
  ScrollArea, users can do that externally now