Diff algorithm speedup
The previous version of the algorithm was almost a line for line copy
paste from

https://blog.jcoglan.com/2017/02/17/the-myers-diff-algorithm-part-3/

The idea of that version was to take the shortest edit distance
algorithm from the myers paper, and remember all steps as we went. This
resulted in a couple undesirable properties

* One allocation per "step" of the algorithm
* Each allocation was a snapshot of that "step" of the algorithm
* Duplication of values. Each step only alters odd or even indices of
  the working vector, so each copy is duplicating data
* Extra memory usage. Since the algorithm is based around a single array
  representing the maximum distance (M + N), each step snapshots this
  whole array. However the valid range of the index changes as we get
  farther from the initial state

Rework the algorithm around the idea of the trace. Each step of the
algorithm will not work in a temporary buffer that is copied into a
trace, but will work directly in the trace to avoid extra allocations.

Note that the single allocation will always allocate the maximum memory
needed, so in some cases we may end up using more memory, but in initial
testing this does not seem to be a significant problem. The degenerate
case is much less of a problem now than it was before

Further improvements could be made to stage allocations to avoid
assuming worst case up front