Limit memory usage
Due to the O(n^2) nature of the current myers diff algorithm
implementation, we see in the wild situations that cause OOMs,
application crashes, swaps of death etc. when generating diffs for files
in spit

Set an upper bound on the amount of memory that we are willing to
allocate for a single diff. This results in a failable diff which needs
to be propagated and handled.

* Add a max memory limit to the diff options bar
* When we try to allocate too much, fail with an out of memory error
* Hold diff results instead of diffs. In the case of failed diffs we
  show an error as the diff content, and color the overview red